Fix bug where post-me handshake finishes before methods available
Previously, there would be a short window of time after the post-me handshake
would finish when the methods on the MySky object would not have been available
yet. This manifested in errors such as e.g. "No such method checkLogin" when
trying to call checkLogin.

Also, this commit reverts the variable named "DEV_MODE" to just "dev" for
consistency with the rest of the codebase.